BUILD: quic+h3: 32-bit compilation errors fixes
In GH #1760 (which is marked as being a feature), there were compilation
errors on MacOS which could be reproduced in Linux when building 32-bit code
(-m32 gcc option). Most of them were due to variables types mixing in QUIC_MIN macro
or using size_t type in place of uint64_t type.

Must be backported to 2.6.

(cherry picked from commit 628e89cfaeec20bcb7687185efac91d6eb36ac07)

+++ src/qpack-dec.c
@@ -167,7 +167,7 @@ int qpack_decode_dec(struct qcs *qcs, void *ctx)
  * Return a negative error if failed, 0 if not.
  */
 static int qpack_decode_fs_pfx(uint64_t *enc_ric, uint64_t *db, int *sign_bit,
-                               const unsigned char **raw, size_t *len)
+                               const unsigned char **raw, uint64_t *len)
 {
 	*enc_ric = qpack_get_varint(raw, len, 8);
 	if (*len == (uint64_t)-1)
@@ -189,7 +189,7 @@ static int qpack_decode_fs_pfx(uint64_t *enc_ric, uint64_t *db, int *sign_bit,
  * Returns 0 on success. In case of error, a negative code QPACK_ERR_* is
  * returned.
  */
-int qpack_decode_fs(const unsigned char *raw, size_t len, struct buffer *tmp,
+int qpack_decode_fs(const unsigned char *raw, uint64_t len, struct buffer *tmp,
                     struct http_hdr *list, int list_size)
 {
 	uint64_t enc_ric, db;
